Changing spacer to use the full view inside the MealPlanView or StudyRoomGroupView
