+++ Campus-iOS/MapComponent/View/MapScreenView.swift
@@ -14,7 +14,7 @@ struct MapScreenView: View {
     let screenWidth = UIScreen.main.bounds.width
     let screenHeight = UIScreen.main.bounds.width
     
-    @State var position = PanelPosition.bottom
+    //@State var position = PanelPosition.bottom
     @State var panelHeight = UIScreen.main.bounds.height * 0.8
     
     init(vm: MapViewModel) {
@@ -28,36 +28,51 @@ struct MapScreenView: View {
             MapContentView(vm: self.vm)
             VStack{
                 Spacer()
-                VStack {
-                    PanelContentView(vm: self.vm)
-                        .background(Color.white)
-                        .cornerRadius(10, corners: [.topRight, .topLeft])
-                        .shadow(radius: 10)
-                        .frame(height: panelHeight, alignment: .bottomTrailing)
-                        .onTapGesture { } /// VERY IMPORTANT: This was added in order to prevent the panel of jumping when scrolling inside the ScrollView; This is probably a bug in SwiftUI since the ScrollView should NOT trigger the gesture-Function (see the print()-Functions)!
-                        .gesture(
-                            DragGesture()
-                                .onChanged { value in
-//                                    print(">> This hould only be printed when the panel is dragged NOT when you're scrolling inside the ScrollView! If so the SwfitUI bug still exists and makes the empty .onTapGesture{} above necessary")
-                                    if let newPanelHeight = check(panelHeight - value.translation.height) {
-                                            panelHeight = newPanelHeight
-                                    }
-                                }
-                                .onEnded { _ in
-//                                    print("<< Now dragging has finished")
+                ZStack {
+                    VStack {
+                        PanelContentView(vm: self.vm)
+                            .background(Color.white)
+                            .cornerRadius(10, corners: [.topRight, .topLeft])
+                            .shadow(radius: 10)
+                            .frame(height: panelHeight, alignment: .bottomTrailing)
+                            .task(id: vm.panelPos) {
+                                if panelHeight != vm.panelPos.rawValue {
                                     withAnimation(.interpolatingSpring(stiffness: 300.0, damping: 30.0, initialVelocity: 10.0)) {
-                                        snapPanel(from: panelHeight)
+                                        panelHeight = vm.panelPos.rawValue
+                                        print(panelHeight)
                                     }
-                                }
-                        )
-                        .task(id: vm.panelPos) {
-                            if panelHeight != vm.panelPos.rawValue {
-                                withAnimation(.interpolatingSpring(stiffness: 300.0, damping: 30.0, initialVelocity: 10.0)) {
-                                    panelHeight = vm.panelPos.rawValue
-                                    print(panelHeight)
                                 }
                             }
+//                            .onTapGesture { } /// VERY IMPORTANT: This was added in order to prevent the panel of jumping when scrolling inside the ScrollView; This is probably a bug in SwiftUI since the ScrollView should NOT trigger the gesture-Function (see the print()-Functions)!
+                    }
+                    
+                    VStack {
+//                        Spacer()
+                        let dragAreaHeight = PanelHeight.top*0.04
+                        ZStack {
+                            Rectangle()
+                                .foregroundColor(.white)
+                            RoundedRectangle(cornerRadius: CGFloat(5.0) / 2.0)
+                                .frame(width: 40, height: CGFloat(5.0))
+                                .foregroundColor(Color.primary.opacity(0.2))
                         }
+                            .cornerRadius(10, corners: [.topLeft, .topRight])
+                            .frame(height: dragAreaHeight, alignment: .top)
+                            .gesture(
+                                DragGesture()
+                                    .onChanged { value in
+                                        if let newPanelHeight = check(panelHeight - value.translation.height) {
+                                            panelHeight = newPanelHeight
+                                        }
+                                    }
+                                    .onEnded { _ in
+                                        withAnimation(.interpolatingSpring(stiffness: 300.0, damping: 30.0, initialVelocity: 10.0)) {
+                                            snapPanel(from: panelHeight)
+                                        }
+                                    }
+                            )
+                        Spacer().frame(height: panelHeight-dragAreaHeight)
+                    }
                 }
             }
         }.edgesIgnoringSafeArea(.top)
