+++ Campus-iOS/MapComponent/View/MapView.swift
@@ -1,27 +1,28 @@
+////
+////  MapView.swift
+////  Campus-iOS
+////
+////  Created by August Wittgenstein on 16.12.21.
+////
 //
-//  MapView.swift
-//  Campus-iOS
+//import SwiftUI
+//import MapKit
 //
-//  Created by August Wittgenstein on 16.12.21.
+//struct MapView: View {
+//    @StateObject var vm: MapViewModel
+//    
+//    var body: some View {
+//        ZStack {
+////            MapContentView(vm: self.vm)
+//            //PanelView(vm: self.vm)
+//            PanelContentView(vm: self.vm)
+//        }
+//        
+//    }
+//}
 //
-
-import SwiftUI
-import MapKit
-
-struct MapView: View {
-    @StateObject var vm: MapViewModel
-    
-    var body: some View {
-        ZStack {
-//            MapContentView(vm: self.vm)
-            PanelView(vm: self.vm)
-        }
-        
-    }
-}
-
-struct MapView_Previews: PreviewProvider {
-    static var previews: some View {
-        MapView(vm: MapViewModel(cafeteriaService: CafeteriasService(), studyRoomsService: StudyRoomsService()))
-    }
-}
+//struct MapView_Previews: PreviewProvider {
+//    static var previews: some View {
+//        MapView(vm: MapViewModel(cafeteriaService: CafeteriasService(), studyRoomsService: StudyRoomsService()))
+//    }
+//}
