+++ Campus-iOS/MapComponent/View/PanelContentCafeteriasListView.swift
@@ -33,9 +33,11 @@ struct PanelContentCafeteriasListView: View {
         List {
             ForEach (self.sortedCafeterias.indices.filter({ searchString.isEmpty ? true : self.sortedCafeterias[$0].name.localizedCaseInsensitiveContains(searchString) }), id: \.self) { id in
                 Button(action: {
+                    print("BEFORE: \(viewModel.selectedCafeteria)")
                     viewModel.selectedCafeteria = self.sortedCafeterias[id]
-                    viewModel.panelPosition = "pushMid"
+                    viewModel.panelPos = .middle
                     viewModel.lockPanel = false
+                    print("AFTER: \(viewModel.selectedCafeteria)")
                 }, label: {
                     CafeteriaRowView(cafeteria: self.$sortedCafeterias[id])
                 })
